# swagger.json
{
    "openapi":  "3.0.0",
    "paths":  {
...
                  "/api/v1/users":  {
                                        "post":  {
                                                     "operationId":  "UsersController_create_v1",
                                                     "parameters":  [

                                                                    ],
                                                     "requestBody":  {
                                                                         "required":  true,
                                                                         "content":  {
                                                                                         "application/json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "$ref":  "#/components/schemas/CreateUserDto"
                                                                                                                             }
                                                                                                              }
                                                                                     }
                                                                     },
                                                     "responses":  {
                                                                       "201":  {
                                                                                   "description":  "User created successfully"
                                                                               }
                                                                   },
                                                     "summary":  "Create a new user",
                                                     "tags":  [
                                                                  "Users"
                                                              ]
                                                 },
                                        "get":  {
                                                    "operationId":  "UsersController_list_v1",
                                                    "parameters":  [

                                                                   ],
                                                    "responses":  {
                                                                      "200":  {
                                                                                  "description":  "Users retrieved successfully"
                                                                              }
                                                                  },
                                                    "summary":  "List users with pagination",
                                                    "tags":  [
                                                                 "Users"
                                                             ]
                                                }
                                    },
                  "/api/v1/users/{id}":  {
                                             "get":  {
                                                         "operationId":  "UsersController_getById_v1",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "id",
                                                                                "required":  true,
                                                                                "in":  "path",
                                                                                "description":  "User ID",
                                                                                "schema":  {
                                                                                               "type":  "string"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "User retrieved successfully"
                                                                                   }
                                                                       },
                                                         "summary":  "Get user by ID",
                                                         "tags":  [
                                                                      "Users"
                                                                  ]
                                                     },
                                             "patch":  {
                                                           "operationId":  "UsersController_update_v1",
                                                           "parameters":  [
                                                                              {
                                                                                  "name":  "id",
                                                                                  "required":  true,
                                                                                  "in":  "path",
                                                                                  "description":  "User ID",
                                                                                  "schema":  {
                                                                                                 "type":  "string"
                                                                                             }
                                                                              }
                                                                          ],
                                                           "requestBody":  {
                                                                               "required":  true,
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/UpdateUserDto"
                                                                                                                                   }
                                                                                                                    }
                                                                                           }
                                                                           },
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "User updated successfully"
                                                                                     }
                                                                         },
                                                           "summary":  "Update user details",
                                                           "tags":  [
                                                                        "Users"
                                                                    ]
                                                       }
                                         },
                  "/api/v1/users/{id}/status":  {
                                                    "patch":  {
                                                                  "operationId":  "UsersController_updateStatus_v1",
                                                                  "parameters":  [
                                                                                     {
                                                                                         "name":  "id",
                                                                                         "required":  true,
                                                                                         "in":  "path",
                                                                                         "description":  "User ID",
                                                                                         "schema":  {
                                                                                                        "type":  "string"
                                                                                                    }
                                                                                     }
                                                                                 ],
                                                                  "requestBody":  {
                                                                                      "required":  true,
                                                                                      "content":  {
                                                                                                      "application/json":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/UpdateUserStatusDto"
                                                                                                                                          }
                                                                                                                           }
                                                                                                  }
                                                                                  },
                                                                  "responses":  {
                                                                                    "200":  {
                                                                                                "description":  "User status updated successfully"
                                                                                            }
                                                                                },
                                                                  "summary":  "Activate or deactivate a user",
                                                                  "tags":  [
                                                                               "Users"
                                                                           ]
                                                              }
                                                },
                  "/api/v1/users/{id}/roles":  {
                                                   "put":  {
                                                               "operationId":  "UsersController_assignRoles_v1",
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "id",
                                                                                      "required":  true,
                                                                                      "in":  "path",
                                                                                      "description":  "The ID of the user to assign roles to",
                                                                                      "schema":  {
                                                                                                     "type":  "string"
                                                                                                 }
                                                                                  }
                                                                              ],
                                                               "requestBody":  {
                                                                                   "required":  true,
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/AssignRolesDto"
                                                                                                                                       }
                                                                                                                        }
                                                                                               }
                                                                               },
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "Roles assigned successfully"
                                                                                         }
                                                                             },
                                                               "summary":  "Assign roles to a user",
                                                               "tags":  [
                                                                            "Users"
                                                                        ]
                                                           }
                                               },
...
              },
    "info":  {
                 "title":  "CRM + HRMS API",
                 "description":  "Production API documentation for the CRM + HRMS platform.",
                 "version":  "1.0.0",
                 "contact":  {

                             }
             },
...
    "servers":  [

                ],
    "components":  {
                       "securitySchemes":  {
                                               "access-token":  {
                                                                    "scheme":  "bearer",
                                                                    "bearerFormat":  "JWT",
                                                                    "type":  "http",
                                                                    "description":  "Enter JWT access token"
                                                                }
                                           },
                       "schemas":  {
...
                                       "CreateUserDto":  {
                                                             "type":  "object",
                                                             "properties":  {
                                                                                "email":  {
                                                                                              "type":  "string",
                                                                                              "description":  "Email address of the user",
                                                                                              "example":  "john.doe@example.com"
                                                                                          },
                                                                                "password":  {
                                                                                                 "type":  "string",
                                                                                                 "description":  "Password for the user",
                                                                                                 "example":  "StrongP@ssw0rd!",
                                                                                                 "minLength":  8
                                                                                             }
                                                                            },
                                                             "required":  [
                                                                              "email",
                                                                              "password"
                                                                          ]
                                                         },
                                       "UpdateUserDto":  {
                                                             "type":  "object",
                                                             "properties":  {
                                                                                "email":  {
                                                                                              "type":  "string",
                                                                                              "description":  "Updated email address of the user",
                                                                                              "example":  "jane.doe@example.com"
                                                                                          }
                                                                            }
                                                         },
                                       "UpdateUserStatusDto":  {
                                                                   "type":  "object",
                                                                   "properties":  {
                                                                                      "status":  {
                                                                                                     "type":  "string",
                                                                                                     "description":  "New status for the user",
                                                                                                     "enum":  [
                                                                                                                  "active",
                                                                                                                  "inactive",
                                                                                                                  "locked"
                                                                                                              ],
                                                                                                     "example":  "inactive"
                                                                                                 }
                                                                                  },
                                                                   "required":  [
                                                                                    "status"
                                                                                ]
                                                               },
                                       "AssignRolesDto":  {
                                                              "type":  "object",
                                                              "properties":  {
                                                                                 "roleIds":  {
                                                                                                 "description":  "Array of role IDs to assign to the user",
                                                                                                 "example":  [
                                                                                                                 "65b1c9e8d4a3e2f1a8b9c0d1"
                                                                                                             ],
                                                                                                 "type":  "array",
                                                                                                 "items":  {
                                                                                                               "type":  "string"
                                                                                                           }
                                                                                             }
                                                                             },
                                                              "required":  [
                                                                               "roleIds"
                                                                           ]
                                                          },
...
                                   }
                   }
}